delete: Modifier 삭제 및 통합
