+++ swiftUIStudy/ButtonMF.swift
@@ -1,0 +1,55 @@
+//
+//  OperationMF.swift
+//  swiftUIStudy
+//
+//  Created by 이지훈 on 5/10/24.
+//
+
+import SwiftUI
+
+struct CalculatorButtonModifier: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .frame(width: 80, height: 80)
+            .background(Color.gray.opacity(0.5))
+            .cornerRadius(60)
+            .foregroundColor(.white)
+            .font(.system(size: 40))
+    }
+}
+
+struct EtcButtonModifier: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .frame(width: 80, height: 80)
+            .background(Color.gray)
+            .cornerRadius(60)
+            .foregroundColor(.white)
+            .font(.system(size: 40))
+    }
+}
+
+struct OperationButtonModifier: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .frame(width: 80, height: 80)
+            .background(Color.orange)
+            .cornerRadius(60)
+            .foregroundColor(.white)
+            .font(.system(size: 40))
+    }
+}
+
+extension View {
+    func calculatorButtonStyle() -> some View {
+        self.modifier(CalculatorButtonModifier())
+    }
+    
+    func etcButtonStyle() -> some View {
+        self.modifier(EtcButtonModifier())
+    }
+
+    func operationButtonStyle() -> some View {
+        self.modifier(OperationButtonModifier())
+    }
+}
